Rename YAML registry loaders to legacy-registry-store.
Move uniqueNames into schemas/common.ts so shared validation is not owned by the retiring registry package.
