+++ packages/server/src/apis/skills.ts
@@ -1,6 +1,6 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
+import type { SkillStore } from '../legacy-registry-store/SkillStore';
 import { listSkillsRoute } from '../routes/skillRoutes';
-import type { SkillStore } from '../store/SkillStore';
 
 export function createSkillsRouter(store: SkillStore) {
   const router = new OpenAPIHono();
