+++ packages/server/src/apis/turns.ts
@@ -20,11 +20,11 @@ import {
 import { streamSSE } from 'hono/streaming';
 import type { Logger } from 'winston';
 import configuration from '../config';
+import type { McpStore } from '../legacy-registry-store/McpStore';
+import type { ModelStore } from '../legacy-registry-store/ModelStore';
 import { createAndExecuteTurnRoute, getTurnRoute, listTurnEventsRoute, listTurnsRoute } from '../routes/turnRoutes';
 import type { ActiveTurnRegistry } from '../runtime/activeTurns';
 import { mintPeeredTurnId } from '../runtime/peeringIds';
-import type { McpStore } from '../store/McpStore';
-import type { ModelStore } from '../store/ModelStore';
 import { TENANT_ID } from './sessions';
 
 function toWireTurn(record: TurnRecordWithoutSnapshot): Turn {
