+++ packages/server/src/legacy-registry-store/McpStore.ts
