+++ packages/server/src/legacy-registry-store/ModelStore.ts
