+++ packages/server/src/main.ts
@@ -31,9 +31,9 @@ try {
     import('./config'),
     import('./db/postgres/client'),
     import('./db/migratePostgres'),
-    import('./store/ModelStore'),
-    import('./store/McpStore'),
-    import('./store/SkillStore'),
+    import('./legacy-registry-store/ModelStore'),
+    import('./legacy-registry-store/McpStore'),
+    import('./legacy-registry-store/SkillStore'),
     import('@truefoundry/utils/agent-session'),
     import('./runtime/activeTurns'),
     import('./runtime/sandboxFactory'),
